Switch to f32

+++ rust/crypto/src/fixedpoint.rs
@@ -1,16 +1,16 @@
 use crate::ring::Ring64Tensor;
 use ndarray::prelude::*;
 
-pub fn ring_encode(x: &ArrayViewD<f64>, scaling_factor: u64) -> Ring64Tensor {
-    let x_upshifted = x * scaling_factor as f64;
+pub fn ring_encode(x: &ArrayViewD<f32>, scaling_factor: u32) -> Ring64Tensor {
+    let x_upshifted = x * scaling_factor as f32;
     let x_converted: ArrayD<i64> = x_upshifted.mapv(|el| el as i64);
     Ring64Tensor::from(x_converted)
 }
 
-pub fn ring_decode(x: &Ring64Tensor, scaling_factor: u64) -> ArrayD<f64> {
+pub fn ring_decode(x: &Ring64Tensor, scaling_factor: u32) -> ArrayD<f32> {
     let x_upshifted: ArrayD<i64> = x.into();
-    let x_converted = x_upshifted.mapv(|el| el as f64);
-    x_converted / scaling_factor as f64
+    let x_converted = x_upshifted.mapv(|el| el as f32);
+    x_converted / scaling_factor as f32
 }
 
 #[cfg(test)]
@@ -23,7 +23,7 @@ mod tests {
             .into_dimensionality::<IxDyn>()
             .unwrap();
 
-        let scaling_factor = 2u64.pow(16);
+        let scaling_factor = 2u32.pow(16);
         let x_encoded = ring_encode(&x.view(), scaling_factor);
         assert_eq!(
             x_encoded,
